Create Age and drink counter
user inputs the age of a person/Persons and if they like soda, popcorn, or both. It then tells you the average age, the youngest age, and the oldest age. it also tells you how many people fall under a age group and what the food preference of everybody is.